nse-bse-bot/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.27
alembic==1.13.1
razorpay==1.3.1
httpx==0.27.0
python-dotenv==1.0.1
bcrypt==4.1.2
email-validator==2.1.0.post1
cors==1.0.1